+++ src/main/java/com/strawberryfarm/fitingle/domain/auth/dto/AuthLoginRequestDto.java
@@ -11,12 +11,12 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Schema(description = "사용자 로그인 요청 Dto")
-public class UsersLoginRequestDto {
+public class AuthLoginRequestDto {
 
-    @Schema(description = "이메일", example = "user@example.com")
-    private String email;
-    @Schema(description = "비밀번호", example = "1234!@")
-    private String password;
-    @Schema(description = "만료 시간", example = "분 단위")
-    private int expiredTime;
+	@Schema(description = "이메일", example = "user@example.com")
+	private String email;
+	@Schema(description = "비밀번호", example = "1234!@")
+	private String password;
+	@Schema(description = "만료 시간", example = "분 단위")
+	private int expiredTime;
 }
