+++ src/main/java/com/strawberryfarm/fitingle/domain/auth/dto/AuthLoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.strawberryfarm.fitingle.domain.users.dto.usersDto;
+package com.strawberryfarm.fitingle.domain.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class UsersLoginResponseDto {
-    private Long userId;
+public class AuthLoginResponseDto {
     private String email;
     private String nickName;
     private String accessToken;
