+++ src/main/java/com/strawberryfarm/fitingle/domain/users/controller/UsersController.java
@@ -1,33 +1,20 @@
 package com.strawberryfarm.fitingle.domain.users.controller;
 
-import com.strawberryfarm.fitingle.domain.users.dto.emailDto.EmailCertificationConfirmRequestDto;
-import com.strawberryfarm.fitingle.domain.users.dto.emailDto.EmailCertificationRequestDto;
 import com.strawberryfarm.fitingle.domain.users.dto.interestArea.InterestAreaRegisterRequestDto;
 import com.strawberryfarm.fitingle.domain.users.dto.keyword.KeywordRegisterRequestDto;
-import com.strawberryfarm.fitingle.domain.users.dto.usersDto.UsersAccessTokenRefreshRequestDto;
 import com.strawberryfarm.fitingle.domain.users.dto.usersDto.UsersDetailUpdateRequestDto;
-import com.strawberryfarm.fitingle.domain.users.dto.usersDto.UsersLoginRequestDto;
-import com.strawberryfarm.fitingle.domain.users.dto.usersDto.UsersLoginResponseVo;
 import com.strawberryfarm.fitingle.domain.users.dto.usersDto.UsersPasswordResetRequestDto;
-import com.strawberryfarm.fitingle.domain.users.dto.usersDto.UsersSignUpRequestDto;
 import com.strawberryfarm.fitingle.domain.users.service.UsersService;
-import com.strawberryfarm.fitingle.dto.ResultDto;
-
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -36,75 +23,18 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(path = "/user",produces = MediaType.APPLICATION_JSON_VALUE)
 @Tag(name = "Users", description = "Users API")
 public class UsersController {
     private final UsersService usersService;
-
-    @PostMapping("/auth/email-request")
-    @Operation(summary = "이메일 인증 요청")
-    public ResponseEntity<?> emailCertificationRequest(HttpServletRequest request, @RequestBody EmailCertificationRequestDto emailCertificationRequestDto) {
-        return ResponseEntity.ok(usersService.emailCertification(emailCertificationRequestDto));
-    }
-
-    @PostMapping("/auth/email-confirm")
-    @Operation(summary = "이메일 인증 확인")
-    public ResponseEntity<?> emailCertificationConfirm(@RequestBody
-    EmailCertificationConfirmRequestDto emailCertificationConfirmRequestDto) {
-        return ResponseEntity.ok(usersService.emailCertificationConfirm(emailCertificationConfirmRequestDto));
-    }
-
-    @PostMapping("/auth/password-edit")
-    @Operation(summary = "비밀 번호 수정 요청")
-    public ResponseEntity<?> passwordEdit(@RequestBody UsersPasswordResetRequestDto usersPasswordResetRequestDto){
-        return ResponseEntity.ok(usersService.passwordEdit(usersPasswordResetRequestDto));
-    }
-
-    @PostMapping("/auth/signup")
-    @Operation(summary = "회원 가입", description = "회원 가입 API 해당 API 호출 전 이메일 인증 과정 필요")
-    public ResponseEntity<?> signUp(@RequestBody UsersSignUpRequestDto usersSignUpRequestDto) {
-        return ResponseEntity.ok(usersService.signUp(usersSignUpRequestDto));
-    }
-
-    @PostMapping("/auth/login")
-    @Operation(summary = "로그인", description = "로그인 api")
-    public ResponseEntity<?> login(@RequestBody UsersLoginRequestDto usersLoginRequestDto, HttpServletResponse httpServletResponse) {
-        ResultDto resultDto = usersService.login(usersLoginRequestDto);
-
-        if (resultDto.getData() == null) {
-            return ResponseEntity.ok(ResultDto.builder()
-                .message(resultDto.getMessage())
-                .data(null)
-                .errorCode(resultDto.getErrorCode())
-                .build());
-        }
-
-        Cookie cookie = new Cookie("refreshToken",((UsersLoginResponseVo)resultDto.getData()).getRefreshToken());
-        cookie.setHttpOnly(true);
-        cookie.setPath("/");
-        httpServletResponse.addCookie(cookie);
-
-        return ResponseEntity.ok(ResultDto.builder()
-            .message(resultDto.getMessage())
-            .data(((UsersLoginResponseVo)resultDto.getData()).getUsersLoginResponseDto())
-            .errorCode(resultDto.getErrorCode())
-            .build());
-    }
-
-    @PostMapping("/auth/logout")
-    @Operation(summary = "로그아웃")
-    public ResponseEntity<?> logout(@CookieValue("refreshToken") String refreshToken) {
-        return ResponseEntity.ok(usersService.logout(refreshToken));
-    }
-
-    @GetMapping("/user")
+    @GetMapping
     @Operation(summary = "유저 상세 정보 가져오기")
     public ResponseEntity<?> getUsersDetail(@AuthenticationPrincipal UserDetails userDetails) {
         Long userId = Long.parseLong(userDetails.getUsername());
         return ResponseEntity.ok(usersService.getUsersDetail(userId));
     }
 
-    @PatchMapping("/user")
+    @PatchMapping
     @Operation(summary = "유저 정보 수정")
     public ResponseEntity<?> updateUsersDetail(@AuthenticationPrincipal UserDetails userDetails, @RequestBody
     UsersDetailUpdateRequestDto usersDetailUpdateRequestDto) {
@@ -112,47 +42,39 @@ public class UsersController {
         return ResponseEntity.ok(usersService.updateUsersDetail(userId,usersDetailUpdateRequestDto));
     }
 
-    @GetMapping("/user/interestArea")
+    @GetMapping("/interestArea")
     @Operation(summary = "관심 지역 가져오기")
     public ResponseEntity<?> getInterestArea(@AuthenticationPrincipal UserDetails userDetails) {
         Long userId = Long.parseLong(userDetails.getUsername());
         return ResponseEntity.ok(usersService.getInterestArea(userId));
     }
 
-    @PostMapping("/user/interestArea")
+    @PostMapping("/interestArea")
     @Operation(summary = "관심 지역 등록")
     public ResponseEntity<?> registerInterestArea(@AuthenticationPrincipal UserDetails userDetails,@RequestBody InterestAreaRegisterRequestDto interestAreaRegisterRequestDto) {
         Long userId = Long.parseLong(userDetails.getUsername());
         return ResponseEntity.ok(usersService.registerInterestArea(userId,interestAreaRegisterRequestDto));
     }
 
-    @GetMapping("/user/keyword")
+    @GetMapping("/keyword")
     @Operation(summary = "키워드 리스트 가져오기")
     public ResponseEntity<?> getKeyword(@AuthenticationPrincipal UserDetails userDetails) {
         Long userId = Long.parseLong(userDetails.getUsername());
         return ResponseEntity.ok(usersService.getKeyword(userId));
     }
 
-    @PostMapping("/user/keyword")
+    @PostMapping("/keyword")
     @Operation(summary = "키워드 등록")
     public ResponseEntity<?> registerKeyword(@AuthenticationPrincipal UserDetails userDetails, @RequestBody KeywordRegisterRequestDto keywordRegisterRequestDto) {
         Long userId = Long.parseLong(userDetails.getUsername());
         return ResponseEntity.ok(usersService.registerKeyword(userId,keywordRegisterRequestDto));
     }
 
-    @DeleteMapping("/user/keyword")
+    @DeleteMapping("/keyword")
     @Operation(summary = "키워드 삭제")
     public ResponseEntity<?> deleteKeyword(@AuthenticationPrincipal UserDetails userDetails,@RequestParam Long keywordId) {
         Long userId = Long.parseLong(userDetails.getUsername());
         return ResponseEntity.ok(usersService.deleteKeyword(userId,keywordId));
-    }
-
-    @PostMapping("/auth/accessToken/refresh")
-    @Operation(summary = "엑세스 토큰 갱신")
-    public ResponseEntity<?> deleteKeyword(@AuthenticationPrincipal UserDetails userDetails, @RequestBody
-        UsersAccessTokenRefreshRequestDto usersAccessTokenRefreshRequestDto) {
-        Long userId = Long.parseLong(userDetails.getUsername());
-        return ResponseEntity.ok(usersService.refreshAccessToken(userId,usersAccessTokenRefreshRequestDto));
     }
 
     @GetMapping("/list")
